Refactor SR breakout predictor with modular architecture and simplified design
