+++ src/tactician/__init__.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 from .leverage_sizer import LeverageSizer
 from .ml_tactics_manager import MLTacticsManager
 from .position_sizer import PositionSizer
-from .sr_breakout_predictor_refactored import SRBreakoutPredictor
+from .sr_breakout_predictor import SRBreakoutPredictor
 from .tactician import Tactician, setup_tactician
 
 # src/tactician/__init__.py
